refactor : 중복코드 제거
- mapper 사용
before : DAO interface 구현한 두 클래스에서 mapping 시 중복코드 다수 존재
after : 중복되는 코드부분을 RowMapper 와 PreparedStatementCreator 로 interface 에 default 로 구현하여 중복코드 제거

- ReservationJdbcDao 메소드 분리
before : 각 메소드에서 db 와 connection 을 만들고 닫는 부분의 코드가 중복됨
after : 메소드 분리하여 중복 제거함

- ReservationService.delete 예외처리
before : 삭제하려는 id 가 존재하지 않아도 별도 코멘트 하지 않음
after : 삭제하려는 id 가 존재하지 않으면 예외발생시켜 처리되지 않음을 응답할 수 있도록 함

+++ src/main/java/nextstep/service/ReservationService.java
@@ -45,6 +45,9 @@ public class ReservationService {
     }
 
     public void delete(Long id) {
+        if (reservationDao.findById(id).isEmpty()) {
+            throw new InvalidInputException(ErrorCode.RESERVATION_NOT_FOUND);
+        }
         reservationDao.delete(id);
     }
 }
